refactor: 클래스 분리 LadderGameColumns  클래스 분리
